add postgresql support
This adds PostgreSQL support for the application. We expect many
students will be using PostgreSQL for their final projects.

This includes tests which mock out the database, tests which exercise a
controller, and a load of EcmaScript Module support fun (of which
TypeScript is a major culprit). It does show that we should have our
bases covered now on all-things related to modules, TypeScript, and
testing.