use separate CSS file
Normally, style-loader will (depending on the size of CSS data) bundle
the CSS data into a style tag in the index.html or later after the
entire JS layer has started. This is slower and doesn't benefit from
browser cache. MiniCssExtractPlugin allows that data to be written to an
individual .css file.