secure and document .env
dotenv allows us to inject some environment variables in our local builds
in a persistent manner. The merging we have allows Webpack to inject
environment variables in via a faked process.env - at build time these
variables are substituted into the code. This can be an inherently
insecure thing to do (as seen by our template, which has our sensitive
database connection string). To combat this, an allow-list is used to
indicate which environment variables we want to use.

Somewhat earmarked is the addition a .env.template, which shows what we
want in our .env file as a starting point.