document the webpack configuration
Much of Webpack configuration is simply getting it working and then
never looking at it again until some major update causes some issues.
Until then the configuration is copied from project to project, with
only minor alterations happening.

This commit documents the webpack config so future generations can make
sense of it and understand why certain features are there (so they can
be preserved or trimmed as necessary). Webpack configurations aren't so
much mysterious as they are time consuming to get just-so.