make DATABASE_URL work for everything
It may not have been working at all. I think what's going on here is
that imports are always hoisted to the top of _modules_. Since we're
using modules now, we can't ensure that we import dotenv and configure
it before requiring other things. So we just import dotenv and configure
it (which might be redundant, but should be fine) when the database is
used in any way.