add some TypeScript caveats
The `.js` thing can't really be called out enough, unfortunately. But it
seems to be the least most evil thing. The most evil thing is having to
care about whether something you're importing is a module vs. a file (or
what version of EcmaScript modules are used). `.js` is downright
frustrating since it confounds our tooling (like you can't jump to a
`.js` file that doesn't exist yet). But a frustration/annoyance doesn't
beat simply being unable to use packages because of the syntax they use
to export their functionality.