separate builds for server and ui, prepare heroku
This splits the server and UI into separate builds. This is because we
want Webpack controlling the UI whereas the server is built via tsc.

This prepares us for an eventual deployment, though nothing is deployed
right now.

Now we support running locally with Webpack and ts-node, with the
addition of running the built app via a vanilla `node` invocation. It
does require the `API_PREFIX` environment variable is set for the
deployed instance if it is to work properly.

Some work had to be done with jest to get it to play nice with some of
the server side changes.