add a complex React test
This adds a more complex test to the foos view in the UI. Now the foo
"name" can be edited. Some backend work had to be done to support
updating.

One of the important bits to remember here is that each React render
requires a new `act` block. The assertions must occur in the `waitFor`
block. You want to `await` these calls.

It's a good idea to ensure the test passing really is passing by
changing the `expect` value to be something intentionally broken. This
is because the default behavior for a test is to pass. If the promise is
orphaned somehow, the test may still pass. Jest has some protections
against this, but they are imperfect.

Earmarked for this: The Express JSON body parser is used now, and CSS
reloads on change.